+++ src/plugins/meteor/assets/meteor-deploy-check.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 APPNAME=<%= appName %>
 APP_PATH=/opt/$APPNAME
 IMAGE=mup-<%= appName %>
@@ -9,10 +11,10 @@ HOST=<%= host %>
 cd $APP_PATH
 
 revert_app (){
-  docker logs --tail=100 $APPNAME 1>&2
+  sudo docker logs --tail=100 $APPNAME 1>&2
 
-  if docker image inspect $IMAGE:previous >/dev/null; then
-    docker tag $IMAGE:previous $IMAGE:latest 
+  if sudo docker image inspect $IMAGE:previous >/dev/null; then
+    sudo docker tag $IMAGE:previous $IMAGE:latest 
     sudo bash $START_SCRIPT > /dev/null 2>&1
 
     echo " " 1>&2
@@ -48,7 +50,7 @@ while [[ true ]]; do
     <% if (host) { %> --header "HOST:$HOST" <% } %>  \
     && exit 0 
 
-  if [ "$elaspsed" == "$DEPLOY_CHECK_WAIT_TIME" ]; then
+  if [ "$elaspsed" "==" "$DEPLOY_CHECK_WAIT_TIME" ]; then
     revert_app
     exit 1
   fi