+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.1 - August 23, 2017
+- Add `mup ssh <server>` command
+- Exit code when task list fails is now 1 instead of 0
+- Fix deploying when server's default shell is zsh @thsowers
+- All docker commands are run with `sudo`
+- `mup proxy stop` doesn't require the `proxy` object to be in the config
+- Add option `app.docker.prepareBundle` to enable or disable prepare bundle
+
 ## 1.3.0 - August 22, 2017
 
 **Hooks**